+++ packages/notification-services-controller/src/NotificationServicesPushController/NotificationServicesPushController.ts
@@ -8,7 +8,6 @@ import type { Messenger } from '@metamask/messenger';
 import type { AuthenticationController } from '@metamask/profile-sync-controller';
 import log from 'loglevel';
 
-import type { Types } from '../NotificationServicesController';
 import type { NotificationServicesPushControllerMethodActions } from './NotificationServicesPushController-method-action-types';
 import type { ENV } from './services/endpoints';
 import type { RegToken } from './services/services';
@@ -18,7 +17,7 @@ import {
   deactivatePushNotifications,
   updateLinksAPI,
 } from './services/services';
-import type { PushNotificationEnv } from './types';
+import type { PushAnalyticsPayload, PushNotificationEnv } from './types';
 import type { PushService } from './types/push-service-interface';
 
 const controllerName = 'NotificationServicesPushController';
@@ -59,12 +58,12 @@ export type NotificationServicesPushControllerStateChangeEvent =
 
 export type NotificationServicesPushControllerOnNewNotificationEvent = {
   type: `${typeof controllerName}:onNewNotifications`;
-  payload: [Types.INotification];
+  payload: [PushAnalyticsPayload];
 };
 
 export type NotificationServicesPushControllerPushNotificationClickedEvent = {
   type: `${typeof controllerName}:pushNotificationClicked`;
-  payload: [Types.INotification];
+  payload: [PushAnalyticsPayload];
 };
 
 export type Events =